qwen_vl_integration/requirements.txt
# Core dependencies for Qwen2-VL support
transformers>=4.45.0
torch>=2.0.0
torchvision>=0.15.0
accelerate>=0.21.0
bitsandbytes>=0.41.0  # For 4-bit quantization
Pillow>=10.0.0
pydantic>=2.5.0
qwen-vl-utils  # Required for process_vision_info

# Additional utilities
numpy>=1.24.0
psutil>=5.9.0  # For memory monitoring

# Existing dependencies (from main API)
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
surya-ocr
python-magic==0.4.27